Convert to demos.muvaster.com
This commit is based on a filtered version of e39a9715b96a73cfcf3a4bff0eeeb8f2e1d31c5b, to remove references to large content blobs in the new repository.

Attempts to replay a working demo on top of *this* commit's history will need to reintroduce the 'panos' directory independently.